+++ sqlparser/lib/src/ast/ast.dart
@@ -30,6 +30,7 @@ part 'schema/column_definition.dart';
 part 'schema/table_definition.dart';
 part 'statements/block.dart';
 part 'statements/create_table.dart';
+part 'statements/create_index.dart';
 part 'statements/create_trigger.dart';
 part 'statements/delete.dart';
 part 'statements/insert.dart';