+++ sqlparser/lib/src/ast/statements/statement.dart
@@ -14,9 +14,14 @@ abstract class CrudStatement extends Statement {
 
 /// Interface for statements that have a primary where clause (select, update,
 /// delete).
-abstract class HasWhereClause implements CrudStatement {
+abstract class HasWhereClause extends Statement {
   Expression get where;
 }
 
 /// Marker interface for statements that change the table structure.
 abstract class SchemaStatement extends Statement implements PartOfMoorFile {}
+
+/// Marker interface for schema statements that create a schematic entity.
+abstract class CreatingStatement extends SchemaStatement {
+  String get createdName;
+}